perf(xtoken): persistent IPC buffer + view-only consumer for teacher logits
Three coupled changes to the cross-tokenizer teacher-logits transport:

1. Producer (dtensor_policy_worker_v2): replace per-step stash of
   per-sample handles with a single persistent ``[B_r, T_t, V_t]``
   IPC buffer captured once on first call. Each step ``.copy_()``-s
   fresh logits into the same backing memory behind the same stable
   handle. ``release_ipc_buffer`` becomes a no-op kept for the driver
   contract.

2. Consumer (loss_functions): rebuild the single rank-level handle
   once and slice ``[mb_start:mb_end]`` for the microbatch — drops
   the per-sample ``rebuild_cuda_tensor_from_ipc`` loop + ``torch.stack``
   and avoids an extra allocation. Asserts contiguous monotonic
   ``sample_idx_within_rank`` to fail loudly if any future change
   reorders microbatch samples.

3. Producer (automodel.FullLogitsPostProcessor): drop the
   ``.to(torch.float32)`` cast — teacher is frozen, downstream
   log_softmax/KL kernels upcast internally where they need fp32, and
   shipping native compute dtype (bf16 under autocast) halves the
   persistent IPC buffer footprint.

Together these eliminate the per-step alloc/free cycle the
consumer-must-copy-out rule was working around, halve the buffer's
memory, and remove the consumer-side stack entirely.

+++ nemo_rl/models/automodel/train.py
@@ -1014,7 +1014,11 @@ class FullLogitsPostProcessor:
                 )
             logits = logits.to_local()
 
-        return logits.to(torch.float32)  # [B, S, V_t]
+        # Teacher is frozen (init_optimizer=False) and the consumer does not
+        # backprop into these logits; downstream log_softmax/KL kernels upcast
+        # to fp32 internally where they need it. Ship native compute dtype
+        # (bf16 under autocast) to halve the IPC buffer footprint.
+        return logits  # [B, S, V_t]
 
 
 class ScorePostProcessor:
